+++ view/comment.go
@@ -11,7 +11,7 @@ type CommentHandler struct {
 }
 
 func (h *CommentHandler) Get (r *http.Request) *ResponseData {
-	s2 := getSubPath(r.URL.Path, 2)
+	s2 := getSubPath(r.URL.Path, 3)
 	if s2 != nil && len(*s2) != 0 {
 		id, err := strconv.ParseUint(*s2, 10, 64)
 		if err != nil {
@@ -40,6 +40,11 @@ func (h *CommentHandler) Get (r *http.Request) *ResponseData {
 }
 
 func (h *CommentHandler) Post (r *http.Request, body *RequestData) *ResponseData {
+	u := loginCheck(r)
+	if u == nil {
+		return InitError("You have not login.")
+	}
+	body.Comment.Author = *u
 	err := model.AddComment(&body.Comment)
 	if err != nil {
 		return InitError(fmt.Sprintf("Cannot create comment due to %s.", err.Error()))
@@ -49,7 +54,11 @@ func (h *CommentHandler) Post (r *http.Request, body *RequestData) *ResponseData
 }
 
 func (h *CommentHandler) Put (r *http.Request, body *RequestData) *ResponseData {
-	s2 := getSubPath(r.URL.Path, 2)
+	u := loginCheck(r)
+	if u == nil {
+		return InitError("You have not login.")
+	}
+	s2 := getSubPath(r.URL.Path, 3)
 	if s2 != nil && len(*s2) != 0 {
 		id, err := strconv.ParseInt(*s2, 10, 64)
 		if err != nil {
@@ -60,6 +69,9 @@ func (h *CommentHandler) Put (r *http.Request, body *RequestData) *ResponseData 
 		if a == nil {
 			return InitError(fmt.Sprintf("Comment %d does not exist.", id))
 		}
+		if body.Comment.Author != *u {
+			return InitError("Unauthorized.")
+		}
 		err = model.UpdateComment(&body.Comment)
 		if err != nil {
 			return InitError(fmt.Sprintf("Cannot update comment %d due to %s.", id, err.Error()))
@@ -69,11 +81,22 @@ func (h *CommentHandler) Put (r *http.Request, body *RequestData) *ResponseData 
 }
 
 func (h *CommentHandler) Delete (r *http.Request) *ResponseData {
-	s2 := getSubPath(r.URL.Path, 2)
+	u := loginCheck(r)
+	if u == nil {
+		return InitError("You have not login.")
+	}
+	s2 := getSubPath(r.URL.Path, 3)
 	if s2 != nil && len(*s2) != 0 {
 		id, err := strconv.ParseInt(*s2, 10, 64)
 		if err != nil {
 			return InitError(err.Error())
+		}
+		a := model.FindCommentById(uint(id))
+		if a == nil {
+			return InitError(fmt.Sprintf("Comment %d does not exist.", id))
+		}
+		if a.Author != *u {
+			return InitError("Unauthorized.")
 		}
 		err = model.DeleteComment(uint(id))
 		if err != nil {
@@ -85,5 +108,5 @@ func (h *CommentHandler) Delete (r *http.Request) *ResponseData {
 }
 
 func init() {
-	http.HandleFunc("/comment/", JsonWrapper(&CommentHandler{}))
+	http.HandleFunc("/api/comment/", JsonWrapper(&CommentHandler{}))
 }
